refactor login route
